Do not include logger implementation in aries-vcx
