+++ aries/aries_vcx/Cargo.toml
@@ -78,7 +78,6 @@ backtrace = { optional = true, version = "0.3" }
 
 [dev-dependencies]
 test_utils = { path = "../misc/test_utils" }
-libvcx_logger = { path = "../misc/legacy/libvcx_logger" }
 wallet_migrator = { path = "../misc/wallet_migrator" }
 async-channel = "1.7.1"
 tokio = { version = "1.20", features = ["rt", "macros", "rt-multi-thread"] }
