+++ aries/aries_vcx/tests/test_mysql_wallet.rs
@@ -10,12 +10,10 @@ mod dbtests {
         base_wallet::{did_wallet::DidWallet, BaseWallet, ManageWallet},
         indy::indy_wallet_config::IndyWalletConfig,
     };
-    use libvcx_logger::LibvcxDefaultLogger;
 
     #[tokio::test]
     #[ignore]
     async fn test_mysql_init_issuer_with_mysql_wallet() -> Result<(), Box<dyn Error>> {
-        LibvcxDefaultLogger::init_testing_logger();
         let db_name = format!("mysqltest_{}", uuid::Uuid::new_v4()).replace('-', "_");
         let storage_config = json!({
             "read_host": "localhost",
