--- aries/misc/legacy/libvcx_logger/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "libvcx_logger"
-version.workspace = true
-authors.workspace = true
-description.workspace = true
-license.workspace = true
-edition.workspace = true
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-aries_vcx_core = { path = "../../../aries_vcx_core" }
-chrono = "0.4"
-env_logger = "0.10"
-log = "0.4"
-
-[target.'cfg(target_os = "android")'.dependencies]
-android_logger = "0.13.3"
