+++ aries/misc/test_utils/src/lib.rs
@@ -3,4 +3,5 @@ pub mod mockdata;
 pub mod random;
 #[rustfmt::skip]
 pub mod constants;
+pub mod logger;
 pub mod mock_wallet;
